ci: move tasks to 'from_deps' transforms instead of mult_dep loader
This will allow us to remove some complexity from the app-services repo
as well as allow us to more easily support multiple types of beetmover
tasks without needing extra kinds.
