+++ taskcluster/ci/signing/kind.yml
@@ -2,32 +2,33 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 ---
-loader: app_services_taskgraph.loader.multi_dep:loader
+loader: taskgraph.loader.transform:loader
 
 transforms:
-  - app_services_taskgraph.transforms.multi_dep:transforms
+  - taskgraph.transforms.from_deps
   - app_services_taskgraph.transforms.deps_complete:alerts
-  - app_services_taskgraph.transforms.signing:transforms
-  - taskgraph.transforms.task:transforms
+  - app_services_taskgraph.transforms.signing
+  - taskgraph.transforms.task
 
 kind-dependencies:
   - module-build
 
-primary-dependency: module-build
-group-by: component
-
-task-template:
-  attributes:
-    shipping_phase: promote
-  run-on-tasks-for: [github-release, cron]
-  description: 'Sign release module {}'
-  worker-type: signing
-  worker:
-    cert:
-      by-level:
-        "3":
-          release-signing
-        default:
-          dep-signing
-  alerts:
-    email: a-s-ci-failures@mozilla.com
+tasks:
+  module-build-signing:
+    from-deps:
+      group-by: component
+      copy-attributes: true
+    attributes:
+      shipping_phase: promote
+    run-on-tasks-for: [github-release, cron]
+    description: 'Sign release module {}'
+    worker-type: signing
+    worker:
+      cert:
+        by-level:
+          "3":
+            release-signing
+          default:
+            dep-signing
+    alerts:
+      email: a-s-ci-failures@mozilla.com
